refactor(data_sources): split qem_dashboard into client/incident/dashboard_openqa
Step 7, part 2 of 3: split per plan §3.3. The 523-line single file
contained three top-level classes; each becomes its own module of the
same name (lowercased to fit Python convention):

  mtui/connector/qem_dashboard.py
    -> mtui/data_sources/qem_dashboard/client.py             (QEMDashboardClient + module constants)
    -> mtui/data_sources/qem_dashboard/incident.py           (QEMIncident)
    -> mtui/data_sources/qem_dashboard/dashboard_openqa.py   (DashboardAutoOpenQA)

`__init__.py` re-exports all three so existing
`from mtui.data_sources.qem_dashboard import …` import patterns keep
working. Class definitions are unchanged.

The two module constants shared between QEMDashboardClient (HTTP
timeout) and DashboardAutoOpenQA (future timeout, FAILED_RESULTS set)
live in `client.py` and are pulled into `dashboard_openqa.py` via
`from .client import FAILED_RESULTS, _FUTURE_TIMEOUT`.

Logger string `mtui.connector.qem_dashboard` stays put in both
`client.py` and `dashboard_openqa.py`. `logging.getLogger()` returns
the same instance for the same name, so the log surface (and
`caplog.at_level(logger="mtui.connector.qem_dashboard")` in the
test) keeps working.

Importer rewrites:
  mtui/commands/{export,reloadoqa,simpleset}.py and mtui/types/updateid.py:
    `from ..connector.qem_dashboard import …`
    -> `from ..data_sources.qem_dashboard import …`
  mtui/types/oqaresults.py docstring class ref (cosmetic).

connector/__init__.py is now empty (docstring-only); oqa_search is
the last module under connector/ and moves in the next commit.

Test edits in tests/test_qem_dashboard_connector.py:
  - import switches `mtui.connector.qem_dashboard` -> `mtui.data_sources.qem_dashboard`
  - whitebox attribute access `qem_dashboard.requests` and
    `qem_dashboard._HTTP_TIMEOUT` now goes through the `client` submodule
    (`_qem_client.requests`, `_qem_client._HTTP_TIMEOUT`) since those
    bindings now live there.
  - `monkeypatch.setattr(qem_dashboard, "_FUTURE_TIMEOUT", …)` is
    retargeted at the `dashboard_openqa` submodule whose namespace
    actually holds the binding that `_load_jobs` looks up.
  - `caplog.at_level("WARNING", logger="mtui.connector.qem_dashboard")`
    is unchanged --- logger name is the external contract.

Deviation from §3.3: the plan did not anticipate that splitting a
single module would change which namespace holds an at-module-level
constant that tests monkeypatch. The fix is mechanical (point the
test at the right submodule) and is the same kind of edit the support/
and cli/ commits had to make for `mock.patch()` strings.

Gates: ruff format/check + ty check + pytest (`not slow and not
network`) all green --- 989 passed.
