+++ mtui/data_sources/qem_dashboard/__init__.py
@@ -1,1 +1,7 @@
 """QEM Dashboard API client and incident model."""
+
+from .client import QEMDashboardClient
+from .dashboard_openqa import DashboardAutoOpenQA
+from .incident import QEMIncident
+
+__all__ = ["DashboardAutoOpenQA", "QEMDashboardClient", "QEMIncident"]
