+++ mtui/types/oqaresults.py
@@ -7,7 +7,7 @@ statically typed and string-key typos are surfaced at definition time.
 The :class:`OpenQAResult` Protocol describes the shared structural surface
 of :class:`mtui.data_sources.openqa.standard.AutoOpenQA`,
 :class:`mtui.data_sources.openqa.kernel.KernelOpenQA` and
-:class:`mtui.connector.qem_dashboard.DashboardAutoOpenQA`. The latter does
+:class:`mtui.data_sources.qem_dashboard.DashboardAutoOpenQA`. The latter does
 not inherit from the ``OpenQA`` ABC -- a Protocol matches that duck-typed
 reality without forcing a class hierarchy change.
 """
