+++ mtui/types/updateid.py
@@ -11,8 +11,8 @@ if TYPE_CHECKING:
     from ..cli.prompter import Prompter
 
 from ..cli.term import prompt_user
-from ..connector.qem_dashboard import DashboardAutoOpenQA, QEMIncident
 from ..data_sources.openqa import KernelOpenQA
+from ..data_sources.qem_dashboard import DashboardAutoOpenQA, QEMIncident
 from ..support.config import Config
 from ..support.exceptions import (
     FailedGiteaCallError,
